library: Don't unpack a file if it already exists
This prevents errors when trying to unpack into an existing read-only
file in the channel.

+++ src/library.rs
@@ -29,6 +29,10 @@ impl CrdtPack<'_> for Library {
     fn unpack(vars: &EnvVars, pack: &Library) -> Result<(), Error> {
         for (filename, filedata) in pack.set.iter() {
             let filepath = vars.data.join(&filename);
+            if filepath.is_file() {
+                continue;
+            }
+
             info!("unpacking {}", &filename);
             let mut file = File::create(&filepath)?;
             file.write_all(&filedata)?;
